+++ common/r_edge.c
@@ -19,6 +19,8 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 */
 // r_edge.c
 
+#include <stdint.h>
+
 #include "quakedef.h"
 #include "r_local.h"
 #include "sound.h"
@@ -565,10 +567,10 @@ void R_ScanEdges(void)
    espan_t *basespans;
    espan_t *basespan_p;
    surf_t *s;
-   
+
    basespans = malloc(sizeof(espan_t)*CACHE_PAD_ARRAY(MAXSPANS, espan_t));
    basespan_p = (espan_t *)
-			((long)(basespans + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
+			((uintptr_t)(basespans + CACHE_SIZE - 1) & ~(uintptr_t)(CACHE_SIZE - 1));
    max_span_p = &basespan_p[MAXSPANS - r_refdef.vrect.width];
 
    span_p = basespan_p;
